Add option to use local filesystem for storage
This PR enables Parseable server to accept additional arguments ( drive or s3) to switch between storage.
Local filesystem behaves like an object storage through implementation of ObjectStorage trait. All
methods of LocalFS are implemented using tokio::fs

Configuration provided on runtime is used to select a storage provider. The storage object is handled as
trait objects hence the requirement of ?Sized on some methods. Calls to S3 are replaced with appropriate methods.
