+++ server/Cargo.toml
@@ -25,6 +25,7 @@ object_store = { version = "0.5.1", features = ["aws"] }
 derive_more = "0.99.17"
 env_logger = "0.9.0"
 futures = "0.3"
+fs_extra = "1.2.0"
 http = "0.2.4"
 humantime-serde = "1.1.1"
 lazy_static = "1.4.0"
@@ -33,6 +34,7 @@ num_cpus = "1.0.0"
 os_info = "3.0.7"
 hostname = "0.3"
 rand = "0.8.4"
+relative-path = "1.7.2"
 rustls = "0.20.6"
 rustls-pemfile = "1.0.1"
 rust-flatten-json = "0.2.0"
@@ -43,10 +45,11 @@ serde_json = "^1.0.8"
 sysinfo = "0.26.4"
 thiserror = "1"
 thread-priority = "0.9.2"
-tokio-stream = "0.1.8"
+tokio-stream = { version = "0.1.8", features = ["fs"] }
 tokio = { version = "1.13.1", default-features = false, features = [
     "sync",
     "macros",
+    "fs",
 ] }
 clokwerk = "0.4.0-rc1"
 actix-web-static-files = "4.0"
