+++ server/src/storage.rs
@@ -16,46 +16,125 @@
  *
  */
 
-use crate::alerts::Alerts;
-use crate::metadata::{LOCK_EXPECT, STREAM_INFO};
+use crate::metadata::STREAM_INFO;
 use crate::option::CONFIG;
-use crate::query::Query;
-use crate::stats::Stats;
+
 use crate::storage::file_link::{FileLink, FileTable};
 use crate::utils;
 
-use async_trait::async_trait;
-use chrono::{NaiveDateTime, Timelike, Utc};
-use datafusion::arrow::datatypes::Schema;
+use chrono::{Local, NaiveDateTime, Timelike, Utc};
 use datafusion::arrow::error::ArrowError;
-use datafusion::arrow::ipc::reader::StreamReader;
-use datafusion::datasource::listing::ListingTable;
 use datafusion::execution::runtime_env::RuntimeEnv;
-use datafusion::parquet::arrow::ArrowWriter;
 use datafusion::parquet::errors::ParquetError;
-use datafusion::parquet::file::properties::WriterProperties;
 use lazy_static::lazy_static;
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 
-use std::collections::HashMap;
-use std::fmt::Debug;
-use std::fs::{self, File};
-use std::iter::Iterator;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 
-use self::file_link::CacheState;
+mod file_link;
+mod localfs;
+mod object_storage;
+mod s3;
+
+pub use localfs::{FSConfig, LocalFS};
+pub use object_storage::{ObjectStorage, ObjectStorageProvider};
+pub use s3::{S3Config, S3};
 
 /// local sync interval to move data.records to /tmp dir of that stream.
 /// 60 sec is a reasonable value.
 pub const LOCAL_SYNC_INTERVAL: u64 = 60;
 
-/// duration used to configure prefix in s3 and local disk structure
+/// duration used to configure prefix in objectstore and local disk structure
 /// used for storage. Defaults to 1 min.
 pub const OBJECT_STORE_DATA_GRANULARITY: u32 = (LOCAL_SYNC_INTERVAL as u32) / 60;
 
+// max concurrent request allowed for datafusion object store
+const MAX_OBJECT_STORE_REQUESTS: usize = 1000;
+
+// all the supported permissions
+// const PERMISSIONS_READ: &str = "readonly";
+// const PERMISSIONS_WRITE: &str = "writeonly";
+// const PERMISSIONS_DELETE: &str = "delete";
+// const PERMISSIONS_READ_WRITE: &str = "readwrite";
+const PERMISSIONS_ALL: &str = "all";
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ObjectStoreFormat {
+    pub version: String,
+    #[serde(rename = "objectstore-format")]
+    pub objectstore_format: String,
+    #[serde(rename = "created-at")]
+    pub created_at: String,
+    pub owner: Owner,
+    pub access: Access,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct Owner {
+    pub id: String,
+    pub group: String,
+}
+
+impl Owner {
+    pub fn new(id: String, group: String) -> Self {
+        Self { id, group }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct Access {
+    pub objects: Vec<AccessObject>,
+}
+
+impl Access {
+    pub fn new(objects: Vec<AccessObject>) -> Self {
+        Self { objects }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct AccessObject {
+    pub id: String,
+    pub group: String,
+    pub permissions: Vec<String>,
+}
+
+impl AccessObject {
+    pub fn new(id: String) -> Self {
+        Self {
+            id: id.clone(),
+            group: id,
+            permissions: vec![PERMISSIONS_ALL.to_string()],
+        }
+    }
+}
+
+impl Default for ObjectStoreFormat {
+    fn default() -> Self {
+        Self {
+            version: "v1".to_string(),
+            objectstore_format: "v1".to_string(),
+            created_at: Local::now().to_rfc3339(),
+            owner: Owner::new("".to_string(), "".to_string()),
+            access: Access::new(vec![]),
+        }
+    }
+}
+
+impl ObjectStoreFormat {
+    fn set_id(&mut self, id: String) {
+        self.owner.id.clone_from(&id);
+        self.owner.group = id;
+    }
+    fn set_access(&mut self, access: Vec<AccessObject>) {
+        self.access.objects = access;
+    }
+}
+
 lazy_static! {
     pub static ref CACHED_FILES: Mutex<FileTable<FileLink>> = Mutex::new(FileTable::new());
+    pub static ref STORAGE_RUNTIME: Arc<RuntimeEnv> = CONFIG.storage().get_datafusion_runtime();
 }
 
 impl CACHED_FILES {
@@ -66,144 +145,6 @@ impl CACHED_FILES {
             .into_iter()
             .flat_map(|ref stream_name| StorageDir::new(stream_name).parquet_files().into_iter())
             .for_each(|ref path| table.upsert(path))
-    }
-}
-
-#[async_trait]
-pub trait ObjectStorage: Sync + 'static {
-    fn new() -> Self;
-    async fn check(&self) -> Result<(), ObjectStorageError>;
-    async fn put_schema(
-        &self,
-        stream_name: String,
-        schema: &Schema,
-    ) -> Result<(), ObjectStorageError>;
-    async fn create_stream(&self, stream_name: &str) -> Result<(), ObjectStorageError>;
-    async fn delete_stream(&self, stream_name: &str) -> Result<(), ObjectStorageError>;
-
-    async fn put_alerts(
-        &self,
-        stream_name: &str,
-        alerts: &Alerts,
-    ) -> Result<(), ObjectStorageError>;
-    async fn put_stats(&self, stream_name: &str, stats: &Stats) -> Result<(), ObjectStorageError>;
-    async fn get_schema(&self, stream_name: &str) -> Result<Option<Schema>, ObjectStorageError>;
-    async fn get_alerts(&self, stream_name: &str) -> Result<Alerts, ObjectStorageError>;
-    async fn get_stats(&self, stream_name: &str) -> Result<Stats, ObjectStorageError>;
-    async fn list_streams(&self) -> Result<Vec<LogStream>, ObjectStorageError>;
-    async fn upload_file(&self, key: &str, path: &str) -> Result<(), ObjectStorageError>;
-    fn query_table(&self, query: &Query) -> Result<ListingTable, ObjectStorageError>;
-    fn query_runtime_env(&self) -> Arc<RuntimeEnv>;
-
-    async fn s3_sync(&self) -> Result<(), MoveDataError> {
-        if !Path::new(&CONFIG.parseable.local_disk_path).exists() {
-            return Ok(());
-        }
-
-        let streams = STREAM_INFO.list_streams();
-
-        let mut stream_stats = HashMap::new();
-
-        for stream in &streams {
-            // get dir
-            let dir = StorageDir::new(stream);
-            // walk dir, find all .arrows files and convert to parquet
-
-            let mut arrow_files = dir.arrow_files();
-            // Do not include file which is being written to
-            let hot_file = dir.path_by_current_time();
-            let hot_filename = hot_file.file_name().expect("is a not none filename");
-
-            arrow_files.retain(|file| {
-                !file
-                    .file_name()
-                    .expect("is a not none filename")
-                    .eq(hot_filename)
-            });
-
-            for file in arrow_files {
-                let arrow_file = File::open(&file).map_err(|_| MoveDataError::Open)?;
-                let reader = StreamReader::try_new(arrow_file, None)?;
-                let schema = reader.schema();
-                let records = reader.filter_map(|record| match record {
-                    Ok(record) => Some(record),
-                    Err(e) => {
-                        log::warn!("warning from arrow stream {:?}", e);
-                        None
-                    }
-                });
-
-                let mut parquet_path = file.clone();
-                parquet_path.set_extension("parquet");
-                let mut parquet_table = CACHED_FILES.lock().unwrap();
-                let parquet_file =
-                    fs::File::create(&parquet_path).map_err(|_| MoveDataError::Create)?;
-                parquet_table.upsert(&parquet_path);
-
-                let props = WriterProperties::builder().build();
-                let mut writer = ArrowWriter::try_new(parquet_file, schema, Some(props))?;
-
-                for ref record in records {
-                    writer.write(record)?;
-                }
-
-                writer.close()?;
-
-                fs::remove_file(file).map_err(|_| MoveDataError::Delete)?;
-            }
-
-            for file in dir.parquet_files() {
-                let metadata = CACHED_FILES.lock().unwrap().get_mut(&file).metadata;
-                if metadata != CacheState::Idle {
-                    continue;
-                }
-
-                let filename = file
-                    .file_name()
-                    .expect("only parquet files are returned by iterator")
-                    .to_str()
-                    .expect("filename is valid string");
-                let file_suffix = str::replacen(filename, ".", "/", 3);
-                let s3_path = format!("{}/{}", stream, file_suffix);
-                CACHED_FILES
-                    .lock()
-                    .unwrap()
-                    .get_mut(&file)
-                    .set_metadata(CacheState::Uploading);
-                let _put_parquet_file = self.upload_file(&s3_path, file.to_str().unwrap()).await?;
-                CACHED_FILES
-                    .lock()
-                    .unwrap()
-                    .get_mut(&file)
-                    .set_metadata(CacheState::Uploaded);
-
-                stream_stats
-                    .entry(stream)
-                    .and_modify(|size| *size += file.metadata().map_or(0, |meta| meta.len()))
-                    .or_insert_with(|| file.metadata().map_or(0, |meta| meta.len()));
-
-                CACHED_FILES.lock().unwrap().remove(&file);
-            }
-        }
-
-        for (stream, compressed_size) in stream_stats {
-            let stats = STREAM_INFO
-                .read()
-                .expect(LOCK_EXPECT)
-                .get(stream)
-                .map(|metadata| {
-                    metadata.stats.add_storage_size(compressed_size);
-                    Stats::from(&metadata.stats)
-                });
-
-            if let Some(stats) = stats {
-                if let Err(e) = self.put_stats(stream, &stats).await {
-                    log::warn!("Error updating stats to s3 due to error [{}]", e);
-                }
-            }
-        }
-
-        Ok(())
     }
 }
 
@@ -266,94 +207,6 @@ impl StorageDir {
     }
 }
 
-pub mod file_link {
-    use std::{
-        collections::HashMap,
-        path::{Path, PathBuf},
-    };
-
-    pub trait Link {
-        fn links(&self) -> usize;
-        fn increase_link_count(&mut self) -> usize;
-        fn decreate_link_count(&mut self) -> usize;
-    }
-
-    #[derive(Debug, Default, PartialEq, Eq, Clone, Copy)]
-    pub enum CacheState {
-        #[default]
-        Idle,
-        Uploading,
-        Uploaded,
-    }
-
-    #[derive(Debug)]
-    pub struct FileLink {
-        link: usize,
-        pub metadata: CacheState,
-    }
-
-    impl Default for FileLink {
-        fn default() -> Self {
-            Self {
-                link: 1,
-                metadata: CacheState::Idle,
-            }
-        }
-    }
-
-    impl FileLink {
-        pub fn set_metadata(&mut self, state: CacheState) {
-            self.metadata = state
-        }
-    }
-
-    impl Link for FileLink {
-        fn links(&self) -> usize {
-            self.link
-        }
-
-        fn increase_link_count(&mut self) -> usize {
-            self.link.saturating_add(1)
-        }
-
-        fn decreate_link_count(&mut self) -> usize {
-            self.link.saturating_sub(1)
-        }
-    }
-
-    pub struct FileTable<L: Link + Default> {
-        inner: HashMap<PathBuf, L>,
-    }
-
-    impl<L: Link + Default> FileTable<L> {
-        pub fn new() -> Self {
-            Self {
-                inner: HashMap::default(),
-            }
-        }
-
-        pub fn upsert(&mut self, path: &Path) {
-            if let Some(entry) = self.inner.get_mut(path) {
-                entry.increase_link_count();
-            } else {
-                self.inner.insert(path.to_path_buf(), L::default());
-            }
-        }
-
-        pub fn remove(&mut self, path: &Path) {
-            let Some(link_count) = self.inner.get_mut(path).map(|entry| entry.decreate_link_count()) else { return };
-            if link_count == 0 {
-                let _ = std::fs::remove_file(path);
-                self.inner.remove(path);
-            }
-        }
-
-        pub fn get_mut(&mut self, path: &Path) -> &mut L {
-            self.inner.get_mut(path).unwrap()
-        }
-    }
-}
-
 #[derive(Debug, thiserror::Error)]
 pub enum MoveDataError {
     #[error("Unable to Open file after moving")]
@@ -363,7 +216,7 @@ pub enum MoveDataError {
     #[error("Could not generate parquet file")]
     Parquet(#[from] ParquetError),
     #[error("Object Storage Error {0}")]
-    ObjectStorag(#[from] ObjectStorageError),
+    ObjectStorage(#[from] ObjectStorageError),
     #[error("Could not generate parquet file")]
     Create,
     #[error("Could not delete temp arrow file")]
@@ -372,16 +225,25 @@ pub enum MoveDataError {
 
 #[derive(Debug, thiserror::Error)]
 pub enum ObjectStorageError {
-    #[error("Bucket {0} not found")]
-    NoSuchBucket(String),
+    // no such key inside the object storage
+    #[error("{0} not found")]
+    NoSuchKey(String),
+
+    // Could not connect to object storage
     #[error("Connection Error: {0}")]
     ConnectionError(Box<dyn std::error::Error + Send + 'static>),
+
+    // IO Error when reading a file or listing path
     #[error("IO Error: {0}")]
     IoError(#[from] std::io::Error),
+
+    // Datafusion error during a query
     #[error("DataFusion Error: {0}")]
     DataFusionError(#[from] datafusion::error::DataFusionError),
+
     #[error("Unhandled Error: {0}")]
     UnhandledError(Box<dyn std::error::Error + Send + 'static>),
+
     #[error("Authentication Error: {0}")]
     AuthenticationError(Box<dyn std::error::Error + Send + 'static>),
 }
